fix: refactor prompt for unknown auto mode
